+++ tests/unzip.rs
@@ -117,12 +117,12 @@ async fn entry_ordering(entry_names: HashSet<String>) {
 
 #[proptest(async = "tokio")]
 async fn file_modification_time(
-    #[strategy(1980u16..=2107u16)] year: u16,
-    #[strategy(1u8..=12u8)] month: u8,
-    #[strategy(1u8..=31u8)] day: u8,
-    #[strategy(0u8..=23u8)] hour: u8,
-    #[strategy(0u8..=59u8)] minute: u8,
-    #[strategy(0u8..=59u8)] second: u8,
+    #[strategy(1980..=2107)] year: i32,
+    #[strategy(1u32..=12u32)] month: u32,
+    #[strategy(1u32..=31u32)] day: u32,
+    #[strategy(0u32..=23u32)] hour: u32,
+    #[strategy(0u32..=59u32)] minute: u32,
+    #[strategy(0u32..=59u32)] second: u32,
 ) {
     let mut builder: Builder<()> = Builder::new();
 
@@ -139,12 +139,12 @@ async fn file_modification_time(
 
     let unpacked_timestamp = unpacked.by_index(0).unwrap().last_modified();
 
-    assert!(unpacked_timestamp.year() == year);
-    assert!(unpacked_timestamp.month() == month);
-    assert!(unpacked_timestamp.day() == day);
-    assert!(unpacked_timestamp.hour() == hour);
-    assert!(unpacked_timestamp.minute() == minute);
-    assert!(unpacked_timestamp.second() == second & !1);
+    assert!(unpacked_timestamp.year() == year as u16);
+    assert!(unpacked_timestamp.month() == month as u8);
+    assert!(unpacked_timestamp.day() == day as u8);
+    assert!(unpacked_timestamp.hour() == hour as u8);
+    assert!(unpacked_timestamp.minute() == minute as u8);
+    assert!(unpacked_timestamp.second() == (second & !1) as u8);
 }
 
 async fn build_and_open<T: EntryData>(builder: Builder<T>) -> ZipArchive<std::io::Cursor<Vec<u8>>> {
